+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/bascket/entity/Basket.java
@@ -1,6 +1,11 @@
-package com.sparta.blackwhitedeliverydriver.entity;
+package com.sparta.blackwhitedeliverydriver.bascket.entity;
 
-import com.sparta.blackwhitedeliverydriver.dto.BasketAddRequestDto;
+import com.sparta.blackwhitedeliverydriver.bascket.dto.BasketAddRequestDto;
+import com.sparta.blackwhitedeliverydriver.common.entity.BaseEntity;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderProduct;
+import com.sparta.blackwhitedeliverydriver.product.entity.Product;
+import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.store.entity.Store;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
