+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/bascket/exception/BasketExceptionMessage.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.exception;
+package com.sparta.blackwhitedeliverydriver.bascket.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
